pyproject.toml
[project]
name = "sc-scrobbler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pylast>=5.3.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "soundcloud-v2>=1.6.0",
    "structlog>=24.4.0",
]
